changing PVC ownerReferences from the DV to the VM
ownerReferences on the PVC are changed to match the ownerReferences of
the DataVolume (which would be a reference to parent VM). This allows
for deleting the DataVolume without deleting the PVC and links the PVC
to the VM so the PVC will be cleaned up when the VM is deleted.